ENDPOINT : get task impl

+++ src/todo/application/use_cases.py
@@ -16,7 +16,7 @@ def create_task(
     due_date: Optional[date] = None,
 ) -> Task:
     task = Task(
-        id=0,
+        id=0, # Def par la BDD en auto increment
         title=title,
         description=description,
         due_date=due_date,
@@ -30,5 +30,17 @@ def create_task(
 # =========================
 
 def delete_task(repository: TaskRepository, task_id: int) -> bool:
+    task = repository.get(task_id)
+    if task is None:
+        return False
+
     repository.delete(task_id)
     return True
+
+
+# =========================
+# Récuperation d'une tache
+# =========================
+
+def get_task(repository: TaskRepository, task_id: int) -> Optional[Task]:
+    return repository.get(task_id)
